+++ ProjectSourceCode/index.js
@@ -13,6 +13,8 @@ const bodyParser = require("body-parser");
 const session = require("express-session"); // To set the session object. To store or access session data, use the `req.session`, which is (generally) serialized as JSON by the store.
 const bcrypt = require("bcryptjs"); //  To hash passwords
 const req = require("express/lib/request");
+const axios = require("axios").default;
+
 
 // *****************************************************
 // <!-- Section 2 : Connect to DB -->
@@ -64,7 +66,10 @@ app.use(
     resave: false,
   })
 );
-
+app.use((req,res, next)=> {
+  res.locals.user = req.session.user;
+  next();
+});
 app.use(
   bodyParser.urlencoded({
     extended: true,
@@ -73,15 +78,6 @@ app.use(
 
 app.use("/images", express.static(path.join(__dirname, "images")));
 
-app.get("/login", (req, res) => {
-  res.render("pages/login");
-  //do something
-});
-
-app.get("/register", (req, res) => {
-  res.render("pages/register");
-  //do something
-});
 
 app.get("/tracking", (req, res) => {
   res.render("pages/tracking");
@@ -93,19 +89,158 @@ app.get("/tracking", (req, res) => {
 // *****************************************************
 
 // TODO - Include your API routes here
+//Register
+app.get('/register', (req, res) => {
+  res.render('pages/register.hbs')
+});
 
 app.get("/welcome", (req, res) => {
   res.json({ status: "success", message: "Welcome!" });
 });
+
+
+app.post('/register', async (req, res) =>{
+  let password = req.body.password;
+  let confirmPassword = req.body.confirmPassword;
+  let username = req.body.username;
+  let email = req.body.email;
+  let birthday = req.body.birthday;
+
+  //checks that password and confirm password are the same
+  if (password !== confirmPassword) {
+    return res.render('pages/register',{
+      error:true,
+      message: 'Passwords do not match',
+      username: username,
+      email:email,
+      birthday:birthday
+    });
+  }
+
+  
+  const hash = await bcrypt.hash(req.body.password, 10);
+
+  try {
+    //adds data into user database then redirects user to login page
+    await db.none(`
+      INSERT INTO users (username, hash_password, email, birthday) VALUES ($1, $2, $3, $4);`, [username, hash, email, birthday]);
+      res.status(201).redirect('/login');
+  }
+
+  catch(err) {
+    if (err.code == '23505') {
+      res.render('pages/register',{
+        error:true,
+        message: 'Username already exists',
+      });
+    }
+    else {
+    console.error('error', err);
+    res.redirect('/register');
+    }
+  }
+});
+
+//login
+app.get('/login', (req, res) => {
+  res.render('pages/login.hbs')
+});
+
+app.get("/exercises", async (req, res) => {
+  const options = {
+    method: "GET",
+    url: "https://exercisedb.p.rapidapi.com/exercises?limit=100",
+    headers: {
+      "X-RapidAPI-Key": process.env.API_KEY,
+      "x-rapidapi-host": "exercisedb.p.rapidapi.com",
+    },
+  };
+
+  try {
+    const { data } = await axios.request(options);
+    console.log(data);
+    res.render("pages/exercises.hbs", { data });
+  } catch (error) {
+    console.error(error);
+    res.status(500).render("pages/exercises.hbs", {
+      exercises: [],
+    });
+  }
+});
+
+app.get('/', (req, res) => {
+  res.redirect('/login'); 
+});
+
+app.post('/login', async (req, res) => { 
+  let username = req.body.username;
+
+  try {
+    const user = await db.one(`SELECT * FROM users WHERE username= $1;`, [username]);
+    
+    // check if password from request matches with password in DB
+    const match = await bcrypt.compare(req.body.password, user.hash_password);
+
+    res.status(200);
+
+    if (match) {
+      req.session.user = user;
+      req.session.save();
+
+      res.redirect("/home");
+    }
+
+    else {
+      res.render("pages/login", {
+        message: `Incorrect password`,
+        error: true,
+      });
+    }
+  }
+
+  catch (err) {
+
+    res.status(401);
+    res.redirect("/register")
+  }
+  });
+
+  const auth = (req, res, next) => {
+    if (!req.session.user) {
+      // Default to login page.
+      return res.redirect('/login');
+    }
+    next();
+  };
+  
+  app.use(auth);
 
 app.get("/home", async (req, res) => {
   const today = new Date().toLocaleDateString(); // Get current date
   res.render("pages/home", { date: today });
 });
 
+app.get('/logout', (req, res) => {
+  req.session.destroy(err => {
+    if (err) {
+      return res.render('pages/home', {
+        message: 'Error logging out. Please try again.',
+        error: true, // Indicate an error occurred
+      });
+    }
+    res.render('pages/login', {
+      message: 'Logged out Successfully',
+      error: false,
+    });
+  });
+});
+
+
+
 // *****************************************************
 // <!-- Section 5 : Start Server-->
 // *****************************************************
 // starting the server and keeping the connection open to listen for more requests
+
 module.exports = app.listen(3000);
 console.log("Server is listening on port 3000");